+++ src/index.ts
@@ -4,29 +4,54 @@ import produce from 'immer'
 import { persistCombineReducers, persistStore, Persistor, PersistorOptions, PersistConfig } from 'redux-persist'
 import storage from 'redux-persist/lib/storage'
 
+export interface ImmerOptions {
+  blacklist?: string[]
+  whitelist?: string[]
+}
+
+export interface PersistOptions<S, RS, HSS, ESS> {
+  persistConfig?: Partial<PersistConfig<S, RS, HSS, ESS>>
+  persistorOptions?: PersistorOptions | null
+  callback?: () => any
+}
+
+export interface PluginOptions<S, RS, HSS, ESS> {
+  immerOptions?: ImmerOptions
+  persistOptions?: PersistOptions<S, RS, HSS, ESS>
+}
+
 let persistor: Persistor
 
 export const getPersistor = (): Persistor => persistor
 
-const plugin = <S = any, RS = any, HSS = any, ESS = any>(
-  persistConfig: Partial<PersistConfig<S, RS, HSS, ESS>> = {},
-  persistStoreConfig?: PersistorOptions | null,
-  callback?: () => any
-): Plugin => {
-  const persistMergedConfig: PersistConfig<S, RS, HSS, ESS> = {
+const shouldCombineImmer = (immerOptions: ImmerOptions, key: string): boolean => {
+  const { whitelist, blacklist } = immerOptions
+  if (whitelist && !whitelist.includes(key)) {
+    return false
+  }
+  if (blacklist && blacklist.includes(key)) {
+    return false
+  }
+  return true
+}
+
+const plugin = <S = any, RS = any, HSS = any, ESS = any>(options?: PluginOptions<S, RS, HSS, ESS>): Plugin => {
+  const { immerOptions = {}, persistOptions = {} } = options || {}
+  const persistConfig: PersistConfig<S, RS, HSS, ESS> = {
     key: 'root',
     storage,
-    ...persistConfig
+    ...(persistOptions.persistConfig || {})
   }
+
   return {
     config: {
       redux: {
         combineReducers (reducers) {
-          const reducersWithImmer: ReducersMapObject<any, Action<any>> = {}
+          const handledReducers: ReducersMapObject<any, Action<any>> = {}
           for (const key of Object.keys(reducers)) {
             const reducer = reducers[key]
-            reducersWithImmer[key] = (state, payload) => {
-              if (typeof state === 'object') {
+            handledReducers[key] = (state, payload) => {
+              if (typeof state === 'object' && shouldCombineImmer(immerOptions, key)) {
                 return produce(state, (draft: Models) => {
                   const next = reducer(draft, payload)
                   if (typeof next === 'object') {
@@ -37,12 +62,12 @@ const plugin = <S = any, RS = any, HSS = any, ESS = any>(
               return reducer(state, payload)
             }
           }
-          return persistCombineReducers(persistMergedConfig, reducersWithImmer)
+          return persistCombineReducers(persistConfig, handledReducers)
         }
       }
     },
     onStoreCreated (store) {
-      persistor = persistStore(store, persistStoreConfig, callback)
+      persistor = persistStore(store, persistOptions.persistorOptions, persistOptions.callback)
     }
   }
 }