+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v2.0.0
+
+- feature: can use with `@rematch/loading`
+
+- change: `plugin options` has changed
+
 ## v1.0.3
 
 - change: install redux as devDependencies